Config.uk: Imply LIBUKMMAP / LIBPOSIX_MMAP
Change `select LIBUKMMAP` to `imply LIBUKMMAP`. Do the same for `select
LIBPOSIX_MMAP`. This is because each may not be required for Go builds
(and use the other one instead).

With `select` in place you cannot disable the selection, resulting in a
dependency erorr when enabling the other one.

With `imply` as a soft select, any one can be disabled, removing the
dependency error.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Approved-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #8